+++ crates/cast/tests/e2e/account/create.rs
@@ -6,6 +6,7 @@ use cast::helpers::constants::CREATE_KEYSTORE_PASSWORD_ENV_VAR;
 use indoc::indoc;
 use snapbox::cmd::{cargo_bin, Command};
 use std::{env, fs};
+use tempfile::TempDir;
 use test_case::test_case;
 
 #[tokio::test]
@@ -90,11 +91,8 @@ pub async fn test_happy_case_generate_salt() {
 
 #[tokio::test]
 pub async fn test_happy_case_add_profile() {
-    let current_dir = Utf8PathBuf::from(duplicate_directory_with_salt(
-        CONTRACTS_DIR.to_string() + "/map",
-        "put",
-        "10",
-    ));
+    let current_dir =
+        duplicate_directory_with_salt(CONTRACTS_DIR.to_string() + "/map", "put", "10");
     let accounts_file = "./accounts.json";
 
     let args = vec![
@@ -121,23 +119,20 @@ pub async fn test_happy_case_add_profile() {
         std::str::from_utf8(&out.stdout).expect("failed to convert command output to string");
     assert!(stdout_str.contains("add_profile: Profile successfully added to Scarb.toml"));
 
-    let contents =
-        fs::read_to_string(current_dir.join("Scarb.toml")).expect("Unable to read Scarb.toml");
+    let contents = fs::read_to_string(current_dir.path().join("Scarb.toml"))
+        .expect("Unable to read Scarb.toml");
     assert!(contents.contains("[tool.sncast.my_account]"));
     assert!(contents.contains("account = \"my_account\""));
-
-    fs::remove_dir_all(current_dir).unwrap();
 }
 
 #[tokio::test]
 pub async fn test_happy_case_accounts_file_already_exists() {
-    let current_dir = Utf8PathBuf::from("./tmp-c3");
     let accounts_file = "./accounts.json";
-    fs::create_dir_all(&current_dir).expect("Unable to create directory");
+    let temp_dir = TempDir::new().expect("Unable to create a temporary directory");
 
     fs_extra::file::copy(
         "tests/data/accounts/accounts.json",
-        current_dir.join(accounts_file),
+        temp_dir.path().join(accounts_file),
         &fs_extra::file::CopyOptions::new().overwrite(true),
     )
     .expect("Unable to copy accounts.json");
@@ -158,7 +153,7 @@ pub async fn test_happy_case_accounts_file_already_exists() {
     ];
 
     let snapbox = Command::new(cargo_bin!("sncast"))
-        .current_dir(&current_dir)
+        .current_dir(temp_dir.path())
         .args(args);
     let bdg = snapbox.assert();
     let out = bdg.get_output();
@@ -169,21 +164,19 @@ pub async fn test_happy_case_accounts_file_already_exists() {
     assert!(stdout_str.contains("max_fee: "));
     assert!(stdout_str.contains("address: "));
 
-    let contents =
-        fs::read_to_string(current_dir.join(accounts_file)).expect("Unable to read created file");
+    let contents = fs::read_to_string(temp_dir.path().join(accounts_file))
+        .expect("Unable to read created file");
     assert!(contents.contains("my_account"));
     assert!(contents.contains("deployed"));
-
-    fs::remove_dir_all(current_dir).unwrap();
 }
 
 #[tokio::test]
 pub async fn test_profile_already_exists() {
-    let current_dir = Utf8PathBuf::from(duplicate_directory_with_salt(
+    let current_dir = duplicate_directory_with_salt(
         CONTRACTS_DIR.to_string() + "/constructor_with_params",
         "put",
         "20",
-    ));
+    );
     let accounts_file = "./accounts.json";
 
     let args = vec![
@@ -201,7 +194,7 @@ pub async fn test_profile_already_exists() {
     ];
 
     let snapbox = Command::new(cargo_bin!("sncast"))
-        .current_dir(&current_dir)
+        .current_dir(current_dir.path())
         .args(args);
     let bdg = snapbox.assert();
     let out = bdg.get_output();
@@ -211,8 +204,6 @@ pub async fn test_profile_already_exists() {
     assert!(std_err.contains(
         "error: Failed to add myprofile profile to the Scarb.toml. Profile already exists"
     ));
-
-    fs::remove_dir_all(current_dir).unwrap();
 }
 
 #[tokio::test]
@@ -279,11 +270,8 @@ pub async fn test_happy_case_keystore() {
 
 #[tokio::test]
 pub async fn test_happy_case_keystore_add_profile() {
-    let current_dir = Utf8PathBuf::from(duplicate_directory_with_salt(
-        CONTRACTS_DIR.to_string() + "/map",
-        "put",
-        "50",
-    ));
+    let current_dir =
+        duplicate_directory_with_salt(CONTRACTS_DIR.to_string() + "/map", "put", "50");
     let keystore_path = "my_key.json";
     let account_path = "my_account.json";
     let accounts_file = "accounts.json";
@@ -315,19 +303,19 @@ pub async fn test_happy_case_keystore_add_profile() {
         std::str::from_utf8(&out.stdout).expect("failed to convert command output to string");
     assert!(stdout_str.contains("add_profile: Profile successfully added to Scarb.toml"));
 
-    let contents =
-        fs::read_to_string(current_dir.join("Scarb.toml")).expect("Unable to read Scarb.toml");
+    let contents = fs::read_to_string(current_dir.path().join("Scarb.toml"))
+        .expect("Unable to read Scarb.toml");
     assert!(contents.contains("[tool.sncast.my_account]"));
     assert!(contents.contains("account = \"my_account.json\""));
 
-    let contents =
-        fs::read_to_string(current_dir.join(account_path)).expect("Unable to read created file");
+    let contents = fs::read_to_string(current_dir.path().join(account_path))
+        .expect("Unable to read created file");
     assert!(contents.contains("\"deployment\": {"));
     assert!(contents.contains("\"variant\": {"));
     assert!(contents.contains("\"version\": 1"));
 
-    let contents =
-        fs::read_to_string(current_dir.join("Scarb.toml")).expect("Unable to read Scarb.toml");
+    let contents = fs::read_to_string(current_dir.path().join("Scarb.toml"))
+        .expect("Unable to read Scarb.toml");
     assert!(contents.contains(r#"[tool.sncast.my_account]"#));
     assert!(contents.contains(r#"account = "my_account.json""#));
     assert!(!contents.contains(r#"accounts-file = "accounts.json""#));
@@ -391,7 +379,7 @@ pub fn test_keystore_already_exists(keystore_path: &str, account_path: &str, err
 
 #[tokio::test]
 pub async fn test_happy_case_keystore_int_format() {
-    let keystore_path = "my_key.json";
+    let keystore_path = "my_key_int.json";
     let account_path = "my_account_int.json";
     _ = fs::remove_file(keystore_path);
     _ = fs::remove_file(account_path);
@@ -432,7 +420,7 @@ pub async fn test_happy_case_keystore_int_format() {
 
 #[tokio::test]
 pub async fn test_happy_case_keystore_hex_format() {
-    let keystore_path = "my_key.json";
+    let keystore_path = "my_key_hex.json";
     let account_path = "my_account_hex.json";
     _ = fs::remove_file(keystore_path);
     _ = fs::remove_file(account_path);