+++ crates/cheatnet/src/execution/deprecated/cairo0_execution.rs
@@ -2,7 +2,7 @@ use crate::execution::deprecated::syscalls::CheatableSyscallHandler;
 use crate::state::CheatnetState;
 use blockifier::execution::call_info::CallInfo;
 use blockifier::execution::contract_class::ContractClassV0;
-use blockifier::execution::deprecated_execution::{
+use blockifier::execution::deprecated_entry_point_execution::{
     finalize_execution, initialize_execution_context, prepare_call_arguments, VmExecutionContext,
 };
 use blockifier::execution::entry_point::{
@@ -45,7 +45,7 @@ pub fn execute_entry_point_call_cairo0(
 
     // region: Modified blockifier code
     let mut cheatable_syscall_handler = CheatableSyscallHandler {
-        syscall_handler,
+        child: syscall_handler,
         cheatnet_state,
     };
 
@@ -62,7 +62,7 @@ pub fn execute_entry_point_call_cairo0(
     Ok(finalize_execution(
         vm,
         runner,
-        cheatable_syscall_handler.syscall_handler,
+        cheatable_syscall_handler.child,
         call,
         previous_vm_resources,
         implicit_args,