+++ crates/forge/tests/e2e/forking.rs
@@ -25,7 +25,7 @@ fn without_cache() {
         Collected 1 test(s) from forking package
         Running 1 test(s) from src/
         [PASS] forking::tests::test_fork_simple
-        Tests: 1 passed, 0 failed, 0 skipped
+        Tests: 1 passed, 0 failed, 0 skipped, 0 ignored, 1 filtered out
         "#}
     );
 }
@@ -42,7 +42,12 @@ fn with_cache() {
     );
     let snapbox = test_runner();
 
-    let output = snapbox.current_dir(&temp).assert().code(1);
+    let output = snapbox
+        .current_dir(&temp)
+        .args(["--exact", "forking::tests::test_fork_simple"])
+        .assert()
+        .code(1);
+
     assert_stdout_contains!(
         output,
         indoc! {r#"
@@ -57,7 +62,7 @@ fn with_cache() {
         Failure data:
             original value: [1480335954842313548834020101284630397133856818], converted to a string: [Balance should be 2]
         
-        Tests: 0 passed, 1 failed, 0 skipped
+        Tests: 0 passed, 1 failed, 0 skipped, 0 ignored, 1 filtered out
 
         Failures:
             forking::tests::test_fork_simple
@@ -78,7 +83,12 @@ fn with_clean_cache() {
         .assert()
         .code(0);
 
-    let output = test_runner().current_dir(&temp).assert().code(0);
+    let output = test_runner()
+        .current_dir(&temp)
+        .args(["--exact", "forking::tests::test_fork_simple"])
+        .assert()
+        .code(0);
+
     assert_stdout_contains!(
         output,
         indoc! {r#"
@@ -89,7 +99,37 @@ fn with_clean_cache() {
         Collected 1 test(s) from forking package
         Running 1 test(s) from src/
         [PASS] forking::tests::test_fork_simple
-        Tests: 1 passed, 0 failed, 0 skipped
+        Tests: 1 passed, 0 failed, 0 skipped, 0 ignored, 1 filtered out
+        "#}
+    );
+}
+
+#[test]
+fn printing_latest_block_number() {
+    let temp = setup_package_with_file_patterns(
+        "forking",
+        &[BASE_FILE_PATTERNS, &[&format!("{CACHE_DIR}/*.json")]].concat(),
+    );
+    let snapbox = test_runner();
+
+    let output = snapbox
+        .current_dir(&temp)
+        .args(["--exact", "forking::tests::print_block_number_when_latest"])
+        .assert()
+        .code(0);
+
+    assert_stdout_contains!(
+        output,
+        indoc! {r#"
+        [..]Compiling[..]
+        [..]Finished[..]
+
+
+        Collected 1 test(s) from forking package
+        Running 1 test(s) from src/
+        [PASS] forking::tests::print_block_number_when_latest
+        Number of the block used for fork testing = [..]
+        Tests: 1 passed, 0 failed, 0 skipped, 0 ignored, 1 filtered out
         "#}
     );
 }